Move Export to QueryPage
Ticket: #309
