+++ src/app/Main/Body/QueryPage/Services/ExportServices/StatExportServices/generateStatisticRowsCsvString.service.ts
@@ -1,4 +1,4 @@
-import { modifyTableDataStringService } from "../../../../../Core/modifyTableDataString.service";
+import { modifyTableDataStringService } from "../../../../../../Core/modifyTableDataString.service";
 
 /**
  * @desc Transform the data to a string for the CSV file. Elements of one row are separated by comma.
