+++ src/app/Main/Header/Header.component.tsx
@@ -11,7 +11,6 @@ import {
     bfrPrimaryPalette,
 } from "../../Shared/Style/Style-MainTheme.component";
 import { ZNPaths } from "../../Shared/URLs";
-import { HeaderExportContainerComponent } from "./Export/Header-ExportContainer.component";
 
 const headerStyle = css`
     width: 100%;
@@ -20,15 +19,7 @@ const headerStyle = css`
     flex-direction: column;
     box-sizing: border-box;
 `;
-const subheaderStyle = css`
-    width: 100%;
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    background-color: ${bfrPrimaryPalette[300]};
-    box-sizing: border-box;
-    box-shadow: 0 8px 6px -6px grey;
-`;
+
 const mainHeaderStyle = (open: boolean): SerializedStyles => css`
     width: 100%;
     height: 100%;
@@ -133,12 +124,6 @@ export function HeaderComponent(): JSX.Element {
                     </NavLink>
                 </div>
             </div>
-
-            {queryOpen && (
-                <div css={subheaderStyle}>
-                    <HeaderExportContainerComponent />
-                </div>
-            )}
         </header>
     );
 }
